requirements.txt
psycopg2-binary
python-dotenv
seaborn
matplotlib
pandas
flet
python-dateutil
